Moved loading-companion
