+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/loading-companion/index.tsx
