+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/metacard-quality/presentation.tsx
@@ -16,7 +16,7 @@
 import { hot } from 'react-hot-loader'
 import * as React from 'react'
 import styled from 'styled-components'
-import LoadingCompanion from '../container/loading-companion'
+import LoadingCompanion from '../loading-companion'
 
 type Props = {
   metacardValidation: any
